Update the jrpc2 module to its new location.
The jrpc2 module and subpackages are being rehomed from bitbucket.org to
github.com. This change updates the import paths to the new location, and
updates the dep-vendored copy to the current API at v0.1.0.

Although the previous version was fairly old, this package's use of jrpc2 was
straightforward, so I simply updated the usage sites in place. The resulting
package still builds and passes all its tests.